+++ config/index.d.ts
@@ -1,0 +1,7 @@
+export const coreConfig: ({
+	dtsBanner,
+}:
+	| {
+			dtsBanner?: string;
+	  }
+	| undefined) => ReturnType<typeof import('ttsup').defineConfig>;