+++ docs/src/content/options/legacyTranslate/TranslateGPU.example.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+	import OptionsDemoBase from '$components/options/OptionsDemoBase.svelte';
+</script>
+
+<OptionsDemoBase options={{ legacyTranslate: false, gpuAcceleration: true }}>
+	Modern translate with GPU acceleration
+
+	<svelte:fragment slot="position" let:x let:y>
+		<code>translate: {x}px {y}px 1px</code>
+	</svelte:fragment>
+</OptionsDemoBase>