+++ packages/react/tsup.config.ts
@@ -1,11 +1,3 @@
-import { defineConfig } from 'ttsup';
+import { coreConfig } from '@neodrag/tsup-config';
 
-export default defineConfig({
-	entry: ['./src/index.ts'],
-	external: ['react'],
-	minify: 'terser',
-	format: 'esm',
-	dts: { resolve: true },
-	clean: true,
-	sourcemap: true,
-});
+export default coreConfig({});