+++ packages/core/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @neodrag/core
 
+## 2.0.0
+
+### Patch Changes
+
+- [#60](https://github.com/PuruVJ/neodrag/pull/60) [`bd831dcc`](https://github.com/PuruVJ/neodrag/commit/bd831dcc101d967b78505acd064cdfcde03b62ff) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Nothing
+
+- [#95](https://github.com/PuruVJ/neodrag/pull/95) [`3c10f6ae`](https://github.com/PuruVJ/neodrag/commit/3c10f6ae377c3e9fc9fea963ea99204a4649806c) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Add `legacyTranslate` option, remove memoization code, align to browser's RAF throttling
+
+- [#97](https://github.com/PuruVJ/neodrag/pull/97) [`9e5c4647`](https://github.com/PuruVJ/neodrag/commit/9e5c46477c7781bc75a57944983434a0c8ceff77) Thanks [@PuruVJ](https://github.com/PuruVJ)! - New output formats
+
+- [`da98e910`](https://github.com/PuruVJ/neodrag/commit/da98e910469d63e53e2462e74196bad3b90ea053) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Expose rootNode and currentNode from events. Remove node and domRect
+
+- [#99](https://github.com/PuruVJ/neodrag/pull/99) [`a1572bce`](https://github.com/PuruVJ/neodrag/commit/a1572bce5186051a5114dd580017a49fc2b3c7fc) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Add transform function
+
+- [`8dd0d88f`](https://github.com/PuruVJ/neodrag/commit/8dd0d88ff0458c0bd6d20e3649371fdf732c9ebb) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Add recomputeBounds option
+
+## 2.0.0-next.6
+
+### Patch Changes
+
+- [#99](https://github.com/PuruVJ/neodrag/pull/99) [`a1572bc`](https://github.com/PuruVJ/neodrag/commit/a1572bce5186051a5114dd580017a49fc2b3c7fc) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Add transform function
+
+## 2.0.0-next.5
+
+### Patch Changes
+
+- [#97](https://github.com/PuruVJ/neodrag/pull/97) [`9e5c464`](https://github.com/PuruVJ/neodrag/commit/9e5c46477c7781bc75a57944983434a0c8ceff77) Thanks [@PuruVJ](https://github.com/PuruVJ)! - New output formats
+
+## 2.0.0-next.4
+
+### Patch Changes
+
+- Expose rootNode and currentNode from events. Remove node and domRect
+
+## 2.0.0-next.3
+
+### Patch Changes
+
+- [#95](https://github.com/PuruVJ/neodrag/pull/95) [`0f513db2`](https://github.com/PuruVJ/neodrag/commit/0f513db2c0a88ed03f0472311a03b6ae0e4f9483) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Add `legacyTranslate` option, remove memoization code, align to browser's RAF throttling
+
+## 2.0.0-next.2
+
+### Patch Changes
+
+- [#92](https://github.com/PuruVJ/neodrag/pull/92) [`820307b`](https://github.com/PuruVJ/neodrag/commit/820307b9e2ed5884b2c4d167ba2f7ae2bad14f87) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Add recomputeBounds option
+
+## 2.0.0-next.1
+
+### Patch Changes
+
+- [#90](https://github.com/PuruVJ/neodrag/pull/90) [`8d04e73`](https://github.com/PuruVJ/neodrag/commit/8d04e7327c81ad345610bdc87bcf0f8b6a40fa9e) Thanks [@PuruVJ](https://github.com/PuruVJ)! - Switch to pointer events
+
 ## 2.0.0-next.0
 
 ### Patch Changes